backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
geoalchemy2==0.14.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Async
celery==5.3.6
redis==5.0.1
aiofiles==23.2.1
httpx==0.26.0

# Storage (MinIO/S3)
minio==7.2.3
boto3==1.34.34

# Geospatial
rasterio==1.3.9
pyproj==3.6.1

# Utilities
python-dotenv==1.0.1
